+++ src/Smart.Design.Library.Showcase.WebApp/bundleconfig.json
@@ -3,6 +3,8 @@
     "outputFileName": "wwwroot/SmartLib/js/bundle-client-smart.js",
     "inputFiles": [
       "wwwroot/SmartLib/js/smart-modules/index.js",
+      "wwwroot/SmartLib/js/smart-modules/menu-sublist.js",
+      "wwwroot/SmartLib/js/smart-modules/burger-menu.js"
     ]
   }
 ]
