+++ src/Smart.Design.Library.Showcase.WebApp/wwwroot/SmartLib/css/smart-main.css
@@ -4954,6 +4954,10 @@ c-file-attachment-list
   border-bottom-left-radius: 0.3rem;
 }
 
+c-side-menu__sublist {
+  display: none;
+}
+
 /* ==========================================================================
   c-loader
  ========================================================================== */
