+++ src/Smart.Design.Library.Showcase.WebApp/wwwroot/SmartLib/scss/_c-side-menu.scss
@@ -73,3 +73,7 @@
       border-bottom-left-radius: 0.3rem;
   }
 }
+
+c-side-menu__sublist{
+    display: none;
+}
